pyproject.toml
[project]
name = "artyfy-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin==6.1.0",
    "flask==2.2.3",
    "flask-cors==3.0.10",
    "gunicorn==20.1.0",
    "python-dotenv==1.0.0",
    "werkzeug==2.2.3",
    "sqlalchemy==2.0.9",
    "psycopg2-binary==2.9.5",
    "alembic==1.10.3",
    "pillow==9.5.0",
    "redis==4.5.4",
    "celery==5.2.7",
    "boto3==1.26.115",
    "flask-sqlalchemy==3.0.3",
    "marshmallow==3.19.0",
    "psycopg2>=2.9.10",
]
